separate loaders from models (domain)
