+++ corehq/warehouse/models/staging.py
@@ -24,7 +24,7 @@ from corehq.warehouse.utils import truncate_records_for_cls
 from dimagi.utils.couch.database import iter_docs
 
 
-class StagingTable(models.Model, WarehouseTable):
+class StagingTable(models.Model):
     batch = models.ForeignKey(
         'Batch',
         on_delete=models.PROTECT,
@@ -32,16 +32,6 @@ class StagingTable(models.Model, WarehouseTable):
 
     class Meta(object):
         abstract = True
-
-    @classmethod
-    def commit(cls, batch):
-        cls.clear_records()
-        cls.load(batch)
-        return True
-
-    @classmethod
-    def clear_records(cls):
-        truncate_records_for_cls(cls, cascade=False)
 
 
 class LocationStagingTable(StagingTable, HQToWarehouseETLMixin):
@@ -217,14 +207,12 @@ class UserStagingTable(StagingTable, HQToWarehouseETLMixin):
         return iter_docs(CouchUser.get_db(), user_ids)
 
 
-class DomainStagingTable(StagingTable, HQToWarehouseETLMixin):
-    '''
+class DomainStagingTable(StagingTable):
+    """
     Represents the staging table to dump data before loading into the DomainDim
 
     Grain: domain_id
-    '''
-    slug = DOMAIN_STAGING_SLUG
-
+    """
     domain_id = models.CharField(max_length=255)
     domain = models.CharField(max_length=100)
     default_timezone = models.CharField(max_length=255)
@@ -243,39 +231,6 @@ class DomainStagingTable(StagingTable, HQToWarehouseETLMixin):
 
     domain_last_modified = models.DateTimeField(null=True)
     domain_created_on = models.DateTimeField(null=True)
-
-    @classmethod
-    def dependencies(cls):
-        return []
-
-    @classmethod
-    def field_mapping(cls):
-        return [
-            ('_id', 'domain_id'),
-            ('name', 'domain'),
-            ('default_timezone', 'default_timezone'),
-            ('hr_name', 'hr_name'),
-            ('creating_user_id', 'creating_user_id'),
-            ('project_type', 'project_type'),
-
-            ('is_active', 'is_active'),
-            ('case_sharing', 'case_sharing'),
-            ('commtrack_enabled', 'commtrack_enabled'),
-            ('is_test', 'is_test'),
-            ('location_restriction_for_users', 'location_restriction_for_users'),
-            ('use_sql_backend', 'use_sql_backend'),
-            ('first_domain_for_user', 'first_domain_for_user'),
-
-            ('last_modified', 'domain_last_modified'),
-            ('date_created', 'domain_created_on'),
-            ('doc_type', 'doc_type'),
-        ]
-
-    @classmethod
-    def record_iter(cls, start_datetime, end_datetime):
-        domain_ids = get_domain_ids_by_last_modified(start_datetime, end_datetime)
-
-        return iter_docs(Domain.get_db(), domain_ids)
 
 
 class FormStagingTable(StagingTable, HQToWarehouseETLMixin):
